Rewrote handle_input logic. Add zoom on mouse scroll and drag/not drag by pressing control. and small changes
handle_input does not throw away all actions of one type except the last one. Mouse movements during one tick are 'added' to each other.
Add mouse_scroll ActionType but it is only need for the zoom.
Add ability to zoom in/out on mouse scroll. During the zoom and drag zoom and shift are calculated so that board takes up all screen.
Add transformation reset on R or if file is op
After marking/unmarking stones as dead new gamestate is considred different and are added to the history.
Draw lesser polygons due to not drawing crosses on stones if not needed.
Changed C -> X for marking/unmarking dead groups as groups are marked with crosses.